refactor: move fault to own module
